# Conflicts:
#   README.md
#   data/db.json
#   src/app/app-routing.module.ts
#   src/app/app.module.ts
#   src/app/catalogue/catalogue/catalogue.component.html
#   src/app/catalogue/catalogue/catalogue.component.ts
#   src/app/core/core.module.ts
#   src/app/icons/icons.module.ts
#   src/app/products/models/product.ts
#   src/app/products/pages/page-detail-product/page-detail-product.component.html
#   src/app/products/services/products.service.ts

+++ src/app/users/pages/page-list-users/page-list-users.component.ts
@@ -30,6 +30,7 @@ export class PageListUsersComponent implements OnInit {
   }
 
   delete(user: User) {
+    if(confirm("Voulez vous supprimer l'utilisateur ?"))
     this.usersService.deleteUserById(user.id)
     .subscribe(() => {
       this.getData();
